finish NavBar component
