+++ src/components/pages/home/RenderHomePage.jsx
@@ -1,64 +1,10 @@
 import React from 'react';
-import { Layout, Menu } from 'antd';
-import {
-  HomeOutlined,
-  LogoutOutlined,
-  SettingFilled,
-  FormOutlined,
-  EditOutlined,
-} from '@ant-design/icons';
-import {
-  CREATE_PROGRAM_PATH,
-  EDIT_PROGRAM_PATH,
-  SETTINGS_PATH,
-} from '../../../routes';
-import { Link } from 'react-router-dom';
-
-// this will make for easier routing later on. We can just toss in a "path" as a prop to MenuItem
-// and voilá, it'll be wrapped in a link.
-const MenuItem = ({ key, to, icon, children, ...props }) => {
-  if (to) {
-    return (
-      <Link to={to}>
-        <Menu.Item key={key} icon={icon} {...props}>
-          {children}
-        </Menu.Item>
-      </Link>
-    );
-  }
-  return (
-    <Menu.Item key={key} icon={icon} {...props}>
-      {children}
-    </Menu.Item>
-  );
-};
+import { Layout } from 'antd';
+import { NavBar } from '../navbar';
 
 function RenderHomePage(props) {
   const { userInfo, authService } = props;
-  const { Header, Content, Footer, Sider } = Layout;
-
-  const handleClick = e => {
-    //Checks on the Menu event.key and directs user to appropriate page
-    switch (e.key) {
-      case '1':
-        alert('Dashboard Clicked');
-        break;
-      case '2':
-        alert('Create Program Clicked');
-        break;
-      case '3':
-        alert('Modify Program Clicked');
-        break;
-      case '4':
-        alert('Settings Clicked');
-        break;
-      case '5':
-        authService.logout();
-        break;
-      default:
-        break;
-    }
-  };
+  const { Header, Content, Footer } = Layout;
 
   return (
     <Layout>
@@ -73,37 +19,7 @@ function RenderHomePage(props) {
       </Header>
 
       <Layout>
-        {/*Sidebar Navigation Container*/}
-        <Sider
-          breakpoint="sm"
-          collapsedWidth="0"
-          // onBreakpoint={broken => { }}
-          // onCollapse={(collapsed, type) => { }}
-        >
-          {/*Menu Items*/}
-          <Menu
-            theme="light"
-            mode="inline"
-            defaultSelectedKeys={['1']}
-            onClick={handleClick}
-          >
-            <MenuItem key="1" icon={<HomeOutlined />}>
-              Dashboard
-            </MenuItem>
-            <MenuItem key="2" icon={<FormOutlined />} to={CREATE_PROGRAM_PATH}>
-              Create A Program
-            </MenuItem>
-            <MenuItem key="3" icon={<EditOutlined />} to={EDIT_PROGRAM_PATH}>
-              Modify Program
-            </MenuItem>
-            <MenuItem key="4" icon={<SettingFilled />} to={SETTINGS_PATH}>
-              Settings
-            </MenuItem>
-            <MenuItem key="5" icon={<LogoutOutlined />}>
-              Logout
-            </MenuItem>
-          </Menu>
-        </Sider>
+        <NavBar fireLogout={authService.logout} />
 
         {/*Main Content Area*/}
         <Content style={{ margin: '24px 16px 0' }}>
