+++ src/components/pages/navbar/index.js
@@ -1,0 +1,1 @@
+export { default as NavBar } from './NavbarContainer';
